+++ e2e/env/env.go
@@ -444,8 +444,9 @@ func E2ETests(env e2e.TestEnv) testhelper.Tests {
 		"environment manipulation": c.singularityEnv,
 		"environment option":       c.singularityEnvOption,
 		"environment file":         c.singularityEnvFile,
-		"issue 5057":               c.issue5057, // https://github.com/sylabs/hpcng/issues/5057
-		"issue 5426":               c.issue5426, // https://github.com/sylabs/hpcng/issues/5426
+		"issue 5057":               c.issue5057, // https://github.com/hpcng/issues/5057
+		"issue 5426":               c.issue5426, // https://github.com/hpcng/issues/5426
 		"issue 43":                 c.issue43,   // https://github.com/sylabs/singularity/issues/43
+		"issue 274":                c.issue274,  // https://github.com/sylabs/singularity/issues/274
 	}
 }